Scraper/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
attrs=23.2.0=pypi_0
blinker=1.8.2=pypi_0
bzip2=1.0.8=h2bbff1b_6
ca-certificates=2024.3.11=haa95532_0
certifi=2024.6.2=pypi_0
cffi=1.16.0=pypi_0
charset-normalizer=3.3.2=pypi_0
click=8.1.7=pypi_0
colorama=0.4.6=pypi_0
expat=2.6.2=hd77b12b_0
flask=3.0.3=pypi_0
flask-cors=4.0.1=pypi_0
h11=0.14.0=pypi_0
idna=3.7=pypi_0
itsdangerous=2.2.0=pypi_0
jinja2=3.1.4=pypi_0
libffi=3.4.4=hd77b12b_1
markupsafe=2.1.5=pypi_0
openssl=3.0.13=h2bbff1b_2
outcome=1.3.0.post0=pypi_0
packaging=24.0=pypi_0
pip=24.0=py312haa95532_0
pycparser=2.22=pypi_0
pysocks=1.7.1=pypi_0
python=3.12.3=h1d929f7_1
python-dotenv=1.0.1=pypi_0
requests=2.32.3=pypi_0
selenium=4.21.0=pypi_0
setuptools=69.5.1=py312haa95532_0
sniffio=1.3.1=pypi_0
sortedcontainers=2.4.0=pypi_0
sqlite=3.45.3=h2bbff1b_0
tk=8.6.14=h0416ee5_0
trio=0.25.1=pypi_0
trio-websocket=0.11.1=pypi_0
typing-extensions=4.12.1=pypi_0
tzdata=2024a=h04d1e81_0
urllib3=2.2.1=pypi_0
vc=14.2=h2eaa2aa_1
vs2015_runtime=14.29.30133=h43f2093_3
webdriver-manager=4.0.1=pypi_0
werkzeug=3.0.3=pypi_0
wheel=0.43.0=py312haa95532_0
wsproto=1.2.0=pypi_0
xz=5.4.6=h8cc25b3_1
zlib=1.2.13=h8cc25b3_1
